@@ -2,16 +2,14 @@
 
 declare(strict_types=1);
 
-namespace SixtyEightPublishers\ImageBundle\Event;
+namespace SixtyEightPublishers\FileBundle\Event;
 
-use Nette;
-use Symfony;
+use Nette\Http\FileUpload;
+use Symfony\Contracts\EventDispatcher\Event;
 
-final class FileUploadEvent extends Symfony\Contracts\EventDispatcher\Event
+final class FileUploadEvent extends Event
 {
-	use Nette\SmartObject;
-
-	public const NAME = 'image_bundle.file_upload';
+	public const NAME = 'file_bundle.file_upload';
 
 	/** @var \Nette\Http\FileUpload  */
 	private $fileUpload;
@@ -19,7 +17,7 @@ final class FileUploadEvent extends Symfony\Contracts\EventDispatcher\Event
 	/**
 	 * @param \Nette\Http\FileUpload $fileUpload
 	 */
-	public function __construct(Nette\Http\FileUpload $fileUpload)
+	public function __construct(FileUpload $fileUpload)
 	{
 		$this->fileUpload = $fileUpload;
 	}
@@ -27,7 +25,7 @@ final class FileUploadEvent extends Symfony\Contracts\EventDispatcher\Event
 	/**
 	 * @return \Nette\Http\FileUpload
 	 */
-	public function getFileUpload(): Nette\Http\FileUpload
+	public function getFileUpload(): FileUpload
 	{
 		return $this->fileUpload;
 	}
