@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
-namespace SixtyEightPublishers\ImageBundle\Exception;
+namespace SixtyEightPublishers\FileBundle\Exception;
 
-final class InvalidStateException extends \RuntimeException implements IException
+use RuntimeException;
+
+final class InvalidStateException extends RuntimeException implements ExceptionInterface
 {
 }
