@@ -2,16 +2,17 @@
 
 declare(strict_types=1);
 
-namespace SixtyEightPublishers\ImageBundle\Storage\Manipulator\Sortable;
+namespace SixtyEightPublishers\FileBundle\Storage\Manipulator\Sortable;
 
-use SixtyEightPublishers;
+use SixtyEightPublishers\FileBundle\Entity\FileInterface;
+use SixtyEightPublishers\FileBundle\Storage\Options\OptionsInterface;
 
 final class NonPersistentSortableManipulator extends AbstractSortableManipulator
 {
 	/**
 	 * {@inheritDoc}
 	 */
-	public function doSort(SixtyEightPublishers\ImageBundle\Storage\Options\IOptions $options, SixtyEightPublishers\ImageBundle\DoctrineEntity\IImage $sortedImage, ?SixtyEightPublishers\ImageBundle\DoctrineEntity\IImage $previousImage, ?SixtyEightPublishers\ImageBundle\DoctrineEntity\IImage $nextImage): bool
+	public function doSort(OptionsInterface $options, FileInterface $sortedFile, ?FileInterface $previousFile, ?FileInterface $nextFile): bool
 	{
 		return TRUE;
 	}
